requirements.txt
fastapi[all]
uvicorn
sqlalchemy[asyncio]
asyncpg
stripe
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiohttp
python-dotenv
aiosqlite
weasyprint==61.2
Jinja2==3.1.2
web3==6.15.1
aiogram
httpx
google-auth
google-api-python-client
redis
alembic
structlog
sentry-sdk
argon2-cffi
gunicorn
